Pass ArrayBuffer for improved performance (#14)
THREE.js internally creates any *BufferAttribute by invoking the
constructor on the corresponding TypedArray. If the argument is already
a TypedArray, the constructor will copy the data, even if the type is
the same. By passing an ArrayBuffer, the new TypedArray simply points
to the existing buffer.

See the following for details:

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/TypedArray#Syntax