Improve performance by only consuming one sector per frame (#62)
This change introduces a throttle on the number of sectors consumed per frame. This was one of the main current visual performance issues while rotating the view. Each consumed sector would trigger Three.js to start uploading GPU data and wait for the result, which would take up to a second for some frames.