Make dist folder when running `npm run build`
And introduce run-script-os to handle this on all operating systems.
This is required to be able to build after deleting
`examples/package-lock.json`.

Also update package-lock.json for both `viewer` and `examples`.

Note: Pointing to `../viewer/dist` is necessary for the examples because
TypeScript cannot find the type definitions for
`@cognite/reveal/threejs` otherwise, because they cannot be listed in
`viewer/package.json`.