Add support and example for picking other Three.js objects (#199)
This changes our picking implementation to return an object similar to what Raycaster returns in Three.js. In addition, it changes the API to take a list of CadNode objects for picking that will be rendered individually. This leaves the distance comparison up to the user, but at least allows us to render to depth and tree index buffers without interference from other objects.