Add script to verify examples/package-lock.json
We have had a few instances of PRs that include an
examples/package-lock.json that removes all dependencies listed
under its @cognite/reveal dependency. This change adds a script that
checks that this has not happened and runs it in Travis CI to make sure
we can avoid the conflict resultions that occur because of this problem.