chore: simplify publish scripts (#688)
I branched from the `use-yarn` branch to avoid conflicts. 

I checked the result from running `npm publish --dry-run` and files are the same as it was, except our LICENSE file will be published from now. Currently, it's not. 

So with these changes publishing should be the same as it's now in terms of commands. Simply type `cd viewer && yarn publish` (or npm instead of yarn, shouldn't be any difference). 

I'd like to learn more about releases later. It would be nice to have github releases with changelogs instead of dummy version strings. But that's not something I planned for this PR.