feat(Cognite3dViewer): add `forceRerender(): void` (#686)
This works for me to implement edit rotation in the console. So I can do the following:

```ts
model.updateTransformation(rotationMatrix);
viewer.forceRerender();
```

An alternative implementation I think would be to have render trigger right in `updateTransformation` function. Which in turn should do something like

```ts
// for CognitePointCloudModel
  updateTransformation(matrix: Matrix4): void {
    this.applyMatrix4(matrix);
    this.updateMatrixWorld(false);
    this.potreeNode.requestRedraw()
  }

// for Cognite3dModel
  updateTransformation(matrix: Matrix4): void {
    this.cadNode.applyMatrix4(matrix);
    this.cadNode.updateMatrixWorld(false);
   this.cadNode.requestRedraw()
  }
```

But I just thought maybe it would be convenient to have that forceRedraw fn for some other cases. So left like that.