Add generators for easy package creation (#64)
Add some [plop]-based generators to aid with package creation. The goal
here is to guide the user through all of the boilerplate around adding a
new package to the `@cognite` scope on NPM -- lerna, yarn workspaces,
tsconfig, test harness, etc.

This also moves the packages under `packages/browser/` into
subdirectories for easier organization.

If more template types wish to be added in the future, then they can be.
However, right now we're structuring it with a `_template` folder
sitting alongside the packages themselves for ease of maintenance &
visibility.