Support multiple versioning systems (#123)
Bring support for both single-branch and multi-branch versioning
strategies to the Jenkinsfile. This comes with an overhaul to the
Jenkinsfile pipeline structure to allow the logic to be reused between
the different execution contexts.

This also brings in some modernization of common conventions to the
build process and app structure, such as automatically configuring the
mixpanel metrics.