Provide a mock for @cognite/metrics (#66)
The `@cognite/metrics` package is stubbed out all over the place.
Instead of requiring every app provide their own stubs, provide a simple
mocks override.

This could be extended in the future to provide inspection of the
metrics which were logged and whatnot. However, that can easily come in
a future release.