Wrap the whole body in try/catch (#75)
Be even more fault-tolerant by wrapping the entire body in a the
try/catch. This will prevent a crash where it's possible that `this` is
undefined and reading `this.startProps` will crash. This is situation
represents a bad state in the app, but we still shouldn't crash.

This also introduces some safety around the `I18nContainer`. If the
`@cognite/metrics` library had not yet been initialized, then this might
cause a crash.

Additionally, it moves `@cognite/metrics` to be a peer dependency of
`@cognite/react-errors`. This will prevent different versions from being
used in the applications.