Tidy up eslint rules across packages (#77)
We're long-overdue to not run eslint against all of our packages.
Introduce eslint to most of our packages and clean up the warnings &
errors which were generated. Note that some packages did not get the
eslint treatment (ex: `@cognite/griff-loader`). This is because of the
fact that fixing the eslint warnings would be such an invasive overhaul
that it would mean basically rewriting the package. In the interest of
scoping this change, they have been left out.

Also introduce a rule to forbid the use of importing from
`styled-components/macro`. This import should never be used in
libraries, only in applications.