feat(Cognite3DViewer): add model now returns any supported model type (#904)
It should a bit simplify the usage of Cognite3DViewer when type doesn't really matter. Also, now `addModel` truly adds any model and we have explicitly declared addCadModel as we have for addPointCloudModel. 

I actually think that the reveal manager should have the same API instead of an overloading addModel.