Eslint plugin html referrer (#101)
We are currently having an open security issue across all the applications at Cognite related to too broad referrer policy. We’ve already tried to solve that generically on the FAS level (#tmp-id_token) which failed due to FAS using CDN which does not support setting headers on resources behind load balancer (#incident-202008121531-inspiring-kowalevski).

The only option we ended up with is setting meta tag on the html level for each app. Adding a rule for that