Add a <model-viewer> example (#996)
Adds a `<model-viewer>` example to the documentation. I'm not sure if we want this example to be in the documentation (since it isn't really documenting anything Reveal-specific), but it was an easy environment to make the POC in. Currently the sidebar does not link to the page, but it is available at `reveal/docs/examples/preview`. If we want to keep it in the docs we can update the PR to link from the sidebar if needed. 
I can also move this to a place that would be more appropriate, or we can just leave the branch as a POC for the future.