Switch from domainName to appId and repo (#331)
This is the beginning of the process of migrating from passing `domainName` to passing `appId`/`repo` instead. The problem with domains is that they are a subject of change, while `appId` are not. Also `repo` will most probably be needed for the future migration of configuration data from Firestore to ConfigMaps.