[D3M-4] feat: add UI for reprocessing of old 3d revisions (#145)
For 3d revisions that don't have outputs for the reveal library, on revision details page (`/:tenant/3d-models/:modelId/revisions/:revisionId`), show "Reprocess" button. It will open modal window with a message about reprocessing. 

If the revision is newer than 01.01.2020 it can be reprocessed without creating a new revision. For the older revisions, in case of reprocessing, a new revision from the existing one will be created.

![image](https://user-images.githubusercontent.com/4989157/98678365-b970af80-235e-11eb-939e-4b51d6222b72.png)