[D3M-21] mn/use-react-query (#202)
What's done in terms of refactoring:

* Now the [react-query](https://react-query.tanstack.com/docs/overview) is used for data fetching instead of redux. I decided to not remove redux completely and keep it for a couple of client-side state fields (we store expanded rows in the models-table and also filters and sorting for that table). It is possible to replace those remnants with a simple react context, but I think we might need redux in the future, however, if not - I'll remove it someday later. Everything else is reimplemented with react-query hooks. 
* Most of the app components are rewritten from class-components into functional components with hooks. 

That's basically it. I haven't added any new features here, except maybe some that react-query brings, e.g. data refetching when browser window regains focus (after you do alt+tab for example). So app now should behave more accurately in cases when someone else works with the same data (adds or removes something) - you'll get an update faster because data fetching happens more often. It's not like it was a goal or something, but it just works out of the box so I think it's ok to keep it. 

The motivation to move from redux to react-query was basically the weirdness of the global state (no types, a lot of unused things, redux boilerplate).