Use cogs date range and add time picker (#238)
* Use Cogs' DateRange

* Add and use new TimeSelector component

* Change to inline eslint-disable

Having eslint-disable inline instead of global lets us see clearly where
do we need to change it in the future.

* Remove external lib class dependency

Just copied the styles into the component itself so it doesn't depend on
the react-date-range class name.

* Remove text before the date picker

* Fix default select value

When the value did not correspond to any of the select options, it was
impossible to select the previous value again. So, this adds an empty
hidden option to be selected when the input value doesn't fit any
option.