fix(TreeView): adapt to orderless nodes api (#312)
Unfortunately, the latest API changes about sort order didn't help in our case. We still need that **depth** index on top of models data for requests to be quick. 

Ordered or not, when we request children of some node with **nodeId**, a full table scan must happen to find all the records with the required **depth**. And this is the slow part. Though I don't know why subsequent requests work faster, perhaps some DB optimisations, temporary indices or caching.