Filter charts on frontend (#275)
* Filter charts on frontend

Currently we filter charts to fetch only the ones the user owns in our
fetch from firebase. This PR instead fetches all charts for a tenant,
and filters the list in the ChartList component to show only user owned
charts. This sets us up to display both user owned charts and public
tenant charts in the ChartList component. This filter will be removed
when the firebase security rules are in place.

* Remove user from ChartService

* Fix test by adding mock env