Changing the logic of the proceed button. (#158)
WIP.
Seems like the dispatching I added is interfering with file pickers ability to reset the status to ready for upload. And the  react_devtools_backend.js:2430 Warning: Can't perform a React state update on an unmounted component. This is a no-op, but it indicates a memory leak in your application. To fix, cancel all subscriptions and asynchronous tasks in a useEffect cleanup function.
    in FileUploader (at UploadStep.tsx:73)
    in div (created by cdf-vision-subapp__FileUploaderWrapper)
    in cdf-vision-subapp__FileUploaderWrapper (at UploadStep.tsx:72)
    in UploadStep (at LazyWrapper.tsx:12)

Will remove console.log() etc. before I merge