fix(react-container): remove useLoginToCdf and fix client state (#288)
im not sure if this might introduce a breaking change, but I really think exposing a static function for getting the client which is then static also, from this package is bad. I would recommend maintaining our own CogniteClient in the app until we find a better way. Alternatively we could expose useCogniteSDKClient but that doesn't give much. Maybe create a new public hook that uses the useCogniteSDKClient to create the client from sidecar.
EDIT: added back getCogniteSDKClient but added a new setCogniteSDKClient that is used in useCogniteSDKClient. Its still a brutal hack
EDIT 2: use memo and not setState