chore: Opt out of Bazel for local development (#744)
I figure out it is possible to opt out of Bazel for local development while keeping all the go to definitions and retriggering of watchers. Now jest/start/storybook/etc work as they used to in the past. This should improve DX. The downside is there might be issues revealed during CI which are not visible during local dev, but let's see how many of those we have over the time.