feat(sidecar): port list (#930)
# @cognite/sidecar

## 0.1.0 Sept 23 2021

- Change local service detection to be more generic

So i'm thinking about updating our sidecar helper:

it would make it so that our services all required a different port to run on, the benefits would be that we would never have service conflicts locally.

The pros are that it would allow for easier updates.

ps: feel free to suggest fixes here for the duplicate type definition/other quick logic i banged in. the only requirement is keeping the port in a way so that it constrains uniqueness.