(cognuit) Refactor and fixes in data transfers and detail view (#953)
* fix: rename datatype column

Datatype columns was showing with the name of "Business tags".
Renamed it and orderd so it shows first, nxt to the status.
Still need to add the actual business tag in another commit

* fix: Keep business tags as a column in the data transfer table

Business tags need to be included in the possible columns to add
in the table (they were shadowed by the data types) and revision is
not a column so no need to ignore

* chore: add columns to ignore and update column name mappings

Removes from the list columns that either do not have any useful
information or are not relevant in the source/target context.
Improve some column name mappings in both data transfers and configs.

* fix: don't convert everything to date format

The wildcard entry was forcing all numbers, including ids and
revision counts, to be converted to junk dates.
The one in name had no effect, as names are strings

* chore: support all date columns as dates

Add all other date columns (Cognuit create date in source and target,
source system crete and update) so they also support rendering as dates
and filtering on date range

* fix: render tags for appropriate attributes

datatypes should render as normal, while business and status tags
should be the ones rendered as tags

* chore: update date format in data transfers and object details

* chore: refactor fields in translation tab of detail view

Includes Cognuit details (id and revision), external id,
last update date and removes configuration tag (?) in the
details tab. Also groups the fields in more logical rows

* chore: refactor fields in source tab of detail view

Adds id, revision, grouping, last updated, renames some headings
and regroups row in a more logical grouping

* chore: reorder sections so that source and translation tabs are consistent

In the source tab, the steps came after the object metadata,m while in the
translation tab they were on the top of the page. Reordered so they always
come on top. Also applied the same style that was in the translation tab to
the source and made the texts consistent with each other

* chore: lint and replace status tags name

Quality tags -> Status tags