fix(discover): prevent sending unwanted requests for fetching wellbores (#1102)
# What has changed?

- Used already fetched wellbores when working with the Wells SDK V3.

# Any other notes?

There is an issue related to `tablev3`. The wellbores sub-table is not rendered unless a re-render is triggered. I tried switching back to `tablev2` and it doesn't have that problem. Works as expected. I think this is something to do with `tablev3`. But, left it for now since I have to give priority to the new wells SDK integration stuff.

Initial render
![image](https://user-images.githubusercontent.com/84778551/139187982-d14faaa0-f80c-48d1-b867-c74ea19572c2.png)

After a re-render is triggered
![image](https://user-images.githubusercontent.com/84778551/139188060-f72a154b-c0ac-45a4-ab3d-eaab1f39d98e.png)

FYI: @ronald-cognite