feat(cognuit) Include connector instances in projects (#1296)
* chore/Rename and replace unused interfaces

Small clean up to move a mock response (not really from the API) to a
separate file and rename the tree response

* feat/add connector instances to projects and heartbeats

Connector instances are now an entity of the API instead of a property
of the project. This means, in order to refernce a project, the endpoints
for projects, heartbeats, packages and objects have to reference its
source, instance and external id (name).
In this change, we're allowing the endpoints that list projects to still
list projects by source without specifying the instance (to be changed later).
The projects returned by the endpoint will still contain the instance on them.

* chore/include CWP Jira in allowed TODO comments

Allow for the CWP Jira project as valid to have TODO comments linked
without triggering the linter. Also adds the TODO comments so we can
add instances when listing projects and retrieve instance names
dynamically (both in Jira)

* feat/Include project with connector instance in responses

Inlcude the full project response (id + external id + instance + source)
in object, package and project DTOs for the responses that use it
(Additionally remove create and update dates from project, because they
are not yet recorded in the DB or responses)

* chore/rename response objects to match API

Rename some of the response objects to match the names of the API DTOs
and make clear when they retrieve an object (singular) or a list (plural)

* feat/projects with instance in endpoints, filters and state transfers

Update state methods, pages, endpoints and filters so the project is no longer
just a string with name, but an object with name (external id), instance and source

Reverting the shortcut from the other commit, we now have a ProjectResponse dto,
which includes a mandatory ID, and a Project dto, which is used for requests and filters

The project string was removed from the optional columns to add in the data transfers
page, but a task was included in Jira to add the project name and instance instead

* feat/Update tags and packages endpoints

Besides the projects endpoints (already fixed in previous commits), these
are some of the endpoints that had their routes changed in
https://github.com/cognitedata/cognuit-api/pull/304

* chore/group methods according to API tags

Add the methods to the packages corresponding tho the API tags
based on the new rearrangements in
https://github.com/cognitedata/cognuit-api/pull/304
There is no change of logic in ths commit, only regrouping

* chore/prepare support for listing instances

method will be needed for listing projects per instance and for the
dashboard. Just added the API definition, still needs the useQuery part

* fix/do not use abbreviations

Code review fix

Co-authored-by: Deep <70804363+deep-cognite@users.noreply.github.com>

* fix/ Readd the parent directory keys to queries

Code review fix. I accidentally ermoved the keys because didn't understand how they worked, but Deep explained they are used by React Query :)

Co-authored-by: Deep <70804363+deep-cognite@users.noreply.github.com>

* fix/ use timestamp in milliseconds

With the newst API changes, we also included that timestamps for both heartbeats
and also created/updated dates in objects, translations, revisions, steps history
are always in milliseconds. Thus, the conversion is no longer necessary. As a
saefty measure, however, we also add a check that if the value is less than 10
digits long, we assume it to be in seconds an do the conversion.
(API change: https://github.com/cognitedata/cognuit-api/pull/310)

Co-authored-by: Deep <70804363+deep-cognite@users.noreply.github.com>