Feat: add file selector and use it for selecting model files in custom model settings (#494)
* Add file select filter

Based on asset search component in gearbox.

The file extension is used to filter results as well as
 initial list of visible items before typing.  Currently only needed
 for model configuration of custom models. Can be extended
 if we have the need for similar filter in the future.

* Use file filter for specifying custom models

Replaces the input text field with file filter drop down for better user
experience. The list of items shown initially is based on a search with
file extension. When user types the name, new search is performed
and results are filtered so that only files with relevant file extension
is shown.

Note that, we cannot show all relevant files initially since it will
require fetching all files in the project.