feature(pid): PIP-39 Update download/upload to support Connections and boundingboxes (#1355)
- Add support for downloading, and uploading connections
- Add support for downloading boundingBoxes for lines and symbols as well

Created a new function getDiagramInstanceOutputFormat that get additional information we want to output, like boundingboxes, that currently is not stored in the diagramSymbolInstances. With this change we should now be able to plot our annotations elsewhere to verify that what we are doing is correct, and that the pointspace is correct.