feat(pid): Visualize connections (#1497)
# What has changed?

This PR adds a visual marker for each connection. Some remark I've thought about when creating this:
- It is important that the visual marker is behind the original paths in the document (otherwise it looks very cluttered)
- If you have the case when there is a line -> symbol -> line that is connected in a straight line, it is important that the visual marker doens't start and stop on the same spon in the symbol. That is why I've implemented the `getPointsCloserToEachOther` function
- When a line is connected to another line the (extended) intersection between the lines is used. When the intersection is close enough to one of the line end points, the visual marker is extended towards the line mid points. If the intersection doesn't exist, the line midpoints are used.

I've split up the cases of when a connection goes from symbol -> symbol, symbol -> line and line -> line. To get the starting point from the symbol I use mid point (calculated from the bounding box). To get the starting point from the line, the path segment with the shortest mid point distance to the other line/symbol is used.

Now is looks something like this:
![Skjermbilde 2021-12-16 kl  12 09 51](https://user-images.githubusercontent.com/7979907/146361134-01784838-db02-427b-bd3c-8460ba345b8b.png)

# Any other notes?

I went a bit overboard with the refactoring. The main point is that the new folder structure is
```
src
├── geometry
├── matcher
├── pid
└── utils
```
in `pid_tools`. 

**This PR also fixes as a bug with delete connection (it was negated of what is should have been)**