feat(pid): Add scale invariant matcher (#1537)
Link to JIRA:
[NA]

# What has changed?

This makes the matcher scale invariant. Currently is accepts scale between 0.5 and 1.5, but that can easily be tuned. The scaled in showed in the side panel (not sure if this is needed but is can be nice for debugging short/medium term).

Edit: I've tested with scale threshold between 0.8 and 1.2 and that is clearly too small for some documents. How bad is it really that the matcher finds a heat exchanger circle or line circle? In both these cases these would be annotated as a symbol with more paths, right?