feat(pid): Make concept of symbol types (#1590)
Link to JIRA:
[PIP-46](https://cognitedata.atlassian.net/browse/PIP-46)

# What has changed?

This PR adds a drop down for symbol types when creating a new symbol. Some other related improvements are also made:
- the line/symbol instance text in the side panel now looks like a JSON
- the modal is removed

# Any other notes

The way we model `DiagramSymbolInstance` and `DiagramLineInstance` still doens't feel completely right. A line is marked by the `symbolId` which probably is an overuse of that field. In addition I wanted to add a `svgRepresentationIndex` to the `DiagramSymbolInstance` but that field also doens't really make sense for lines.

Note: This will probably have a medium sized merge conflict with #1566