Clean upload modal when closed (#529)
* Remove dead code

* Make upload modal not closeable

Without this it is possible to click outside of the modal  and
close the it while upload is ongoing  and no files will be shown
in e.g.  process page even  when upload is done.
Files are added to state only when upload is clicked explicitly.

* Clean file list in modal when closed

Previously the uploaded files were kept and
shown in upload modal. Deleted files would therefore  also be visible
in the list.  This is because the
`ModalFileUploader` is using react state hook which is not synced with
redux state.