feat(pid): Add symbol types and special case arrow in auto analysis (#1619)
Link to JIRA:
N/a

# What has changed?

This PR adds some more symbols such as `Cap`, `Insolation` and `Arrow`. For arrow the `Auto analysis` is tweaked to *not* interpret arrow symbol instances as regular symbols (now they will not be connected to anything).

# Any other notes?

The way the symbol type is retrieved in `findLinesAndConnections` is somewhat hackish. Would we rather want to also pass the symbols to that function (and do the lookup) or just store the symbol type directly in the symbol instance? I think I prefer the last solution

Edit: To avoid the hackish way to retrieve the symbol type, and still make it easy to get the symbol type for a diagram instance, I've change the `DiagramInstance` types to

```typescript
export interface DiagramInstance {
  type: string;
  pathIds: string[];
  labelIds: string[];
}

export interface DiagramSymbolInstance extends DiagramInstance {
  symbolId: string;
  scale?: number;
}

export interface DiagramLineInstance extends DiagramInstance {
  type: 'Line';
}
```
This change effects much of the code base, hence the large amount of line changed.