feat(pid): Manually label/add line numbers (#1649)
Link to JIRA:
[PIP-70](https://cognitedata.atlassian.net/browse/PIP-70)

# What has changed?

This PR introduces a new tool to manually label line numbers to line or symbol instances. When that tool is selected, a contextual drop down and input field is shown to add new line numbers or select the active line. When a line nummer is active and you click on an instance that instance will be added to the active line (or removed if it already was labeled). The instances that are not labeled are fainted out.

# Any other notes?

And what icon should we use for this? Currently the string icon is used, but there is probably something better! [Possible icons](https://cognitedata.github.io/cogs.js/?path=/docs/tokens-icons--icon)