improve scrolling with mouse (#1685)
This adds detecting what kind of input device the user is using, and changes the behaviour of the `wheel` event based on that. 

Expected behavior: 
- If the user uses a two finger swipe, the stage will pan
- If the user uses a pinching gesture, the stage will zoom
- If the user scrolls using an external mouse, the stage will zoom. 

The browser support for implementing this is surprisingly bad but the implemented approach is what Miro is actually doing (I debugged their source code), so I think it should be fine. We might want to (but not need to) tweak it for other browsers/OSes later on since there doesn't seem to be a unified interface for detecting this.