feat(pid): Infer position, toPosition and toFile for PID file connections (#1755)
Link to JIRA:
[PIP-79](https://cognitedata.atlassian.net/browse/PIP-79)

# What has changed?

Infer `position`, `toPosition` and `toFile` for P&ID file connections. `toPosition` and `toFile` are inferred based on regex expressions and handled in `addOrRemoveLabelToInstance` (which now are moved to `pid-tools/src/utils/diagramInstanceUtils.ts`). 

The `position` is calculated based on if the 'File connection' is placed on the left ('A' column), right ('B' column) or at the top. If it is placed on the left or right side, the closest pidLabel with regex `pidLabel.text.trim().match(/^[0-9]{1,2}$/)` is used. If it is placed at the top `pidLabel.text.trim().match(/^[A-Z]$/)` check if used.

# Any other notes?

Since I only use a trivial check to find the grid labels based thresholds, it can find the `toPosition` inside the file connection (i.e. find 'N' instead of 'P'). That is why there is two different thresholds:
```typescript
const leftColumnThreshold = 0.08;
const leftFileConnectionThreshold = 0.15;
const rightColumnThreshold = 0.92;
const rightFileConnectionThreshold = 0.85;
const topColumnThreshold = 0.05;
const topFileConnectionThreshold = 0.2;
```

<img width="876" alt="Skjermbilde 2022-01-25 kl  13 32 58" src="https://user-images.githubusercontent.com/7979907/150977673-2ae2f9b2-5230-4a07-b9b0-6dba823d6b8e.png">

It would be nice to not be so sensitive towards these numbers.

I've also introduced some more type safety when it comes to symbol types and to check wether an instance is a file connection. Would be nice with feedback on that!