refactor(cognite-ornate): merge drawingLayer into baseLayer (#1799)
This PR removes the `drawingLayer` and replaces it with the `baseLayer`. 

The benefit of having a separate drawing layer would be to improve performance since the layers could be rerendered independently, but in practice, the current implementation is not correct and leads to the majority of shapes being added to the baseLayer anyway and everything being redrawn on most rerenders. 

This has the side effect of making it more difficult to correctly manage the z-axis positioning of the rendered elements. 

AFAICS, there's also not a simple path to allowing documents to be rendered on one layer, and have the drawings rendered on another layer, and still allow the corresponding groups in the different layers to be moved as one group.