Show Error popup and failed status in process table list view
If error occurs for one of the jobs for a file popover will be disabled in the annotation column. And the error icon will contain a tooltip with error for the particular job type. To do this new state was created to keep error messages for failed files in a job.

File status will be "Failed" is all the jobs for the file are failed.