[VIS-669] Handle when app disconnects while processing files and properly show error messages
Previous behavior was to remove the job if an error occurred while fetching the job status. This changes it to only remove the job if the job was not queued in backend.

Since errors sent by SDK for network disconnection and error responses are different that needs to be properly handled to display error message correctly