feat(discover): minor update to reauth (#1833)
Link to JIRA: https://cognitedata.atlassian.net/browse/PP-2518

# What has changed?

Instead of the redirect, let's try this first. I found an issue during re-auth that was caught and hidden, it said the client was already authenticated and cannot be done so again. So making a new client now during this stage.

Related to: https://github.com/cognitedata/applications/pull/1832

The other fixes in this PR are some other minor things, the one of concern is moving `setReAuth` out of the `projectConfig` condition, since that will never exist at this point.