Feat(simconfig): Update Timeseries resampling routine to match SimConnect [SIM-292] (#1884)
Link to JIRA:
[SIM-292 - Update the Timeseries resampling routine to match the Connector](https://cognitedata.atlassian.net/jira/software/projects/SIM/boards/484?selectedIssue=SIM-292)

# What has changed?

This PR aims to fix issues with time series resampling for cases of bad data quality. Additionally, a small refactor of the `Timeseries` class and associated methods was implemented to better accommodate the resampling routine. 

# Any other notes?

This has already been implemented in SimConnect and needs to be updated here as well to avoid having different logics.