chore(discover): Split saved search save button (#1897)
Link to JIRA:
https://cognitedata.atlassian.net/browse/DSY-702

# What has changed?
We are getting rid of the `customSubComponent` and `subComponentPlacement` props in the Input component. This was the only place where it was used, so to prevent breaking changes, removing it from Discover. 

I spoke with @konstantinbaykovdesignteam and he approved on the idea. We're moving the save button to the right of the input field, instead of inside of it. (Pressing `Enter` will still work)

**BEFORE:**
<img width="275" alt="Screenshot 2022-02-09 at 11 10 48" src="https://user-images.githubusercontent.com/73833726/153176911-387a7e00-0b02-451a-a81b-d6ba7da3eaeb.png">

**AFTER:**
<img width="276" alt="Screenshot 2022-02-09 at 11 11 35" src="https://user-images.githubusercontent.com/73833726/153177033-32a3a48e-72d4-4a19-b30c-aee83e75d1c9.png">