feat(pid): Support for equipment symbols with equipment tags (#1855)
Link to JIRA:
[_insert link here_](https://cognitedata.atlassian.net/jira/software/projects/PIP/boards/534?selectedIssue=PIP-88)

# What has changed?
New symbol type 'equipment'. Tag will be added when connecting the appropriate symbol.

Expanding a symbol type in side panel will select all instances of the symbol in svg, and select the addSymbol type. Intended to simplify both seeing all instances, and to add or remove instances missing or wrongly matched.

Clicking on a instance will select it as the current label target, and select the connect label tool