[VIS-658] Custom model settings (#577)
Currently, users can specify a `.tflite` file (manually uploaded to CDF) to run custom models. With the new features in the API, we can handle the model files so that custom models can be specified using their job Ids.

Accordingly, in this work we update (file) selector to populate the list based on items from `/automl/list` endpoint (excluding failed jobs). Since the model jobs also must have a name, we remove the "custom model name" field and instead use the name set by the user when creating the model.

The models are listed when the process page is initially loaded, so that the user won't notice the listing delay (takes between 3-5s to list the models). The new flow is summarized in the video below:

https://user-images.githubusercontent.com/9768053/153379832-ccea5f3e-3d5a-4deb-a563-6330fa6a9c75.mov