feat(pid): Better text bounding box (#1919)
Link to JIRA:
N/A

# What has changed?

The text will now have a 20% extra width (to accommodate potential different text size in PDFs and SVGs). I've also seen a bug in an Isometric were the bonding box was too small. I think that was which way the rotation is. The new method should handle this case, even with a small performance cost.