feat(pid): More robust file connection logic (#1969)
Link to JIRA:
[PIP-105](https://cognitedata.atlassian.net/browse/PIP-105)

# What has changed?

This PR should address three cases that didn't work properly before.

1. The chosen position label should never be inside the file connection
2. The top right and top left file connections was sometimes connected to the labels on the side even when it pointed Up/Down. This should now be fixed
3. Since the logic before only chose the closest label, it sometimes chose an incorrect label. This is now addressed by making sure the angle from the file connection to the label is close to right, up or left depending on isOnRigthSide, isAtTop or isOnLeftSide respectively

# Any other notes?

Added a `BoudingBox` class. Not sure if that's the best name for that class (maybe `Reactangle` or `Box` is better?). I'm also thinking that this class can be used for better connections from label to symbols/lines in Auto Analysis in a follow-up PR