Fix process page crash on custom model select (#607)
When clicking "Add custom model", it seems like its corresponding
handler function isn't immediately called. This makes it so that the
custom model type is selected, but its model settings are not stored in
the process page state which, consequently, results in an out of bounds
error.

This commit fixes the mentioned issue by adding additional security to
the model select by pre-filtering the options that are not found in the
process page state. Accordingly, it won't matter that the custom model
settings are not added to the `availableDetectionModels` first.

Co-authored-by: Marvin Reza <marvinreza@Marvins-MacBook-Pro.local>