architecture(pid): Move to pure javascript architecture (#1881)
# What has changed?

Change to a pure JavaScript architecture. With this change everything feels 10x more snappy since we reduces a lot of the unnecessary re-renders we got with the "React" approach before. A small side benefit is that this approach is framework agnostic which means that we could change to Svelte, Vue or a pure Javascript stack in the future. 

# Any other notes?

Everything should work as before (just a lot more snappy) except the graph explorer mode which I haven't ported yet. 

There is still room for better structure, the different tools can for instances we refactored similar to how Ornate does it.