feat(pid-tools): Connect labels to lines (#2011)
Link to JIRA:
[PIP-111](https://cognitedata.atlassian.net/browse/PIP-111)

# What has changed?

Also connect labels to lines. The same visualizations as for symbol instances is used to show label to line connection. This doesn't always look great since it uses the line midpoint (when #2010 is merged, I can try to fix that visualization)

# Any other notes?

By changing over to `minBy` (from lodash) the performance is actually A LOT better. From 500ms to 80ms in one case 🚅