feat(pid-tools): Improve visibility for what is labeled as one or multiple symbols (#2010)
Link to JIRA:
[PIP-116](https://cognitedata.atlassian.net/browse/PIP-116)

# What has changed?

A bounding box is placed behind the symbol instance in the add symbol tool. This will hopefully make is easier to see when something is annotated as one symbol or as two symbols connected together.

# Any other notes?

Also refactored the `refresh` code path somewhat. I'm very open to other colors/opacities. I thought it was hard to find the correct balance between usefulness and not looking too cluttered


I've added quite a lot of
```typescript
if (!this.svg) {
  throw Error('Calling refresh without SVG');
}
if (!this.pidDocument) {
  throw Error('Calling refresh pidDocument');
}
```
which probably means that this should be refactored differently. Would like to hear suggestions to do this differently as well